Change Config field type

+++ library/src/manager/utils/bounding_box.rs
@@ -3,9 +3,9 @@ use serde::{Serialize, Deserialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct BoundingBox {
     pub name: String,
-    pub x1: f64,
-    pub x2: f64,
-    pub y1: f64,
-    pub y2: f64,
+    pub x1: u32,
+    pub x2: u32,
+    pub y1: u32,
+    pub y2: u32,
     pub confidence: f64
 }
